Changed Trymove and move verification again
Changed trymove to have it accept a pieces vector as a parameter to allow them to be deleted when captured.

Still need the kings to be able to be moved backwards and when they are initialized as kings for them to be on the correct square.
As of right know when the are initialized they are being placed on the square where the piece was before it was moved instead of on the bank rank of the board.